pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-engine"
version = "0.1.0"
description = "Clean, extensible agentic reasoning system with tool calling"
readme = "README.md"
requires-python = ">=3.9"
authors = [
    {name = "Agata Zywot", email = "your.email@example.com"}
]
keywords = ["llm", "agents", "reasoning", "tools", "research"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "vllm>=0.5.0",
    "transformers>=4.40.0",
    "torch>=2.0.0",
    "pyyaml>=6.0",
    "numpy>=1.24.0",
    "requests>=2.28.0",
    "pillow>=9.0.0",
    "openai>=1.0.0",
    "anthropic>=0.25.0",
    "datasets>=2.18.0",
    "wandb>=0.16.0",
    "jinja2>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/agent-engine"
Repository = "https://github.com/yourusername/agent-engine"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
